third_party/blink/web_tests/fast/events/mouseover-button-expected.txt

Test for bugs cr 87473, 64882.

This test verifies that we get the correct button on mouseover events in .which when entering a document.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS successfullyParsed is true

TEST COMPLETE

PASS document.events['A-no-press'].which is 0
PASS document.events['B-left'].which is 1
PASS document.events['C-no-press'].which is 0
PASS document.events['D-middle'].which is 2
